mcp_packet_server/requirements.txt
# MCP Packet Server Requirements
# Core dependencies for production use

# HTTP requests
requests>=2.28.0

# Google API dependencies
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Todoist API
todoist-api-python>=2.1.3

# Environment variable management
python-dotenv>=1.0.0

# Development dependencies
ruff>=0.3.0
mypy>=1.8.0
types-requests>=2.31.0
